refactor: extract to component
