+++ resources/js/features/game-list/components/WantToPlayGamesDataTable/WantToPlayGamesDataTableToolbar.tsx
@@ -1,13 +1,12 @@
 import type { ColumnFiltersState, Table } from '@tanstack/react-table';
-import { RxCross2 } from 'react-icons/rx';
 
-import { BaseButton } from '@/common/components/+vendor/BaseButton';
 import { usePageProps } from '@/common/hooks/usePageProps';
 import { formatNumber } from '@/common/utils/l10n/formatNumber';
 
 import { getAreNonDefaultFiltersSet } from '../../utils/getAreNonDefaultFiltersSet';
 import { DataTableAchievementsPublishedFilter } from '../DataTableAchievementsPublishedFilter';
 import { DataTableFacetedFilter } from '../DataTableFacetedFilter';
+import { DataTableResetFiltersButton } from '../DataTableResetFiltersButton/DataTableResetFiltersButton';
 import { DataTableSearchInput } from '../DataTableSearchInput';
 import { DataTableViewOptions } from '../DataTableViewOptions';
 
@@ -27,14 +26,6 @@ export function WantToPlayGamesDataTableToolbar<TData>({
 
   const currentFilters = table.getState().columnFilters;
   const isFiltered = getAreNonDefaultFiltersSet(currentFilters, defaultColumnFilters);
-
-  const resetFiltersToDefault = () => {
-    if (defaultColumnFilters) {
-      table.setColumnFilters(defaultColumnFilters);
-    } else {
-      table.resetColumnFilters();
-    }
-  };
 
   return (
     <div className="flex w-full flex-col justify-between gap-2 md:flex-row">
@@ -61,14 +52,11 @@ export function WantToPlayGamesDataTableToolbar<TData>({
         ) : null}
 
         {isFiltered ? (
-          <BaseButton
-            variant="ghost"
-            size="sm"
-            onClick={resetFiltersToDefault}
-            className="px-2 text-link lg:px-3"
-          >
-            Reset <RxCross2 className="ml-2 h-4 w-4" />
-          </BaseButton>
+          <DataTableResetFiltersButton
+            table={table}
+            defaultColumnFilters={defaultColumnFilters}
+            tableApiRouteName="api.user-game-list.index"
+          />
         ) : null}
       </div>
